requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
PyPDF2>=3.0.0
python-docx>=1.1.0
markdown>=3.5.0
langchain>=0.1.0
pydantic-settings>=2.1.0
numpy>=1.24.0
